refactor: Recurring network block of codes in server and client Newfile(s):  - Connection.h - Connection Declaration  - Connection.c - Connection Definition Modified:  - sungka.h - Add Include Guards  - client.c - refactor initializing server and connecting the client to the server with init_server() and connect_client_to_server() function defined in connection.c file  - server.c - refactor initializing server and listening to client program  - makefile - include connection.c in compilation since it is now a dependency of server and client
