+++ src/server.c
@@ -3,19 +3,9 @@
 
   Fix Race Condition type bug
 */
+#include "../include/connection.h"
 #include <stdio.h>
 #include <stdlib.h>
-
-#include <sys/types.h>
-#include <sys/socket.h>
-
-#include <netinet/in.h>
-#include <netdb.h>
-
-#include <unistd.h>
-
-void exit_with_error(char *);
-int listen_and_accept_connections(int);
 
 int main(int argc, char *argv[])
 {
@@ -24,13 +14,10 @@ int main(int argc, char *argv[])
 
   // define the server address
   struct sockaddr_in server_address;
-  server_address.sin_family = AF_INET;
-  server_address.sin_port = htons(9888);
-  server_address.sin_addr.s_addr = INADDR_ANY;
+  init_server(&server_address);
   
   // bind the socket to our specified IP and PORT
   bind(server_socket, (struct sockaddr*) &server_address, sizeof(server_address));
-
   int client_socket = listen_and_accept_connections(server_socket);
 
   while (1) {
@@ -60,25 +47,4 @@ int main(int argc, char *argv[])
   close(client_socket);
 
   return 0;
-}
-
-int listen_and_accept_connections(int server_socket) {
-  // listen for connections
-  listen(server_socket, 5);
-  
-  // accept client socket connection requested on server_socket
-  int client_socket;
-  client_socket = accept(server_socket, NULL, NULL);
-
-  if (client_socket < 0)
-    exit_with_error("Client connection failed");
-  else
-    printf("Connection successful!\n");
-  
-  return client_socket;
-}
-
-void exit_with_error(char *message) {
-  printf("%s", message);
-  exit(-1);
 }
